refactor(experimental)!: move vfio_cuda_iommu_map to iommu_map_pa
The module was named after its consumers rather than its service: it installs
caller-supplied physical addresses into whatever IOMMU domain the target
already uses, whatever produced them and whatever the device is.

Moves only, plus the edits without which the tree stops building: the module's
include and obj-m, the wrapper's fallback include, the test's include, and
meson's subdir(). The identifiers follow in the next commit, and where the
wrapper finally belongs is settled in the one after that.
